Correction of precision in JSD and TVD torch versions.

+++ src/BayesCompare/distances.py
@@ -237,9 +237,9 @@ def _jsd_torch(sigma1, sigma2, mu1=None, mu2=None, num_samples=10000, gen=None):
         logdet1 = torch.sum(torch.log(torch.diag(A1)))
         logdet2 = torch.sum(torch.log(torch.diag(A2)))
         # generate random samples from each distribution
-        x10 = torch.randn((k, num_samples), generator=gen)
+        x10 = torch.randn((k, num_samples), generator=gen, dtype=torch.float64)
         x1 = torch.Tensor(A1 @ x10)
-        x20 = torch.randn((k, num_samples), generator=gen)
+        x20 = torch.randn((k, num_samples), generator=gen, dtype=torch.float64)
         x2 = torch.Tensor(A2 @ x20)
         # compute densities for each
         p1 = -torch.sum(x10**2, 0) / 2 - logdet1
@@ -393,9 +393,9 @@ def _tvd_torch(sigma1, sigma2, mu1=None, mu2=None, num_samples=10000, gen=None):
         logdet1 = torch.sum(torch.log(torch.diag(A1)))
         logdet2 = torch.sum(torch.log(torch.diag(A2)))
         # generate random samples from each distribution
-        x10 = torch.randn((k, num_samples), generator=gen)
+        x10 = torch.randn((k, num_samples), generator=gen, dtype=torch.float64)
         x1 = torch.Tensor(A1 @ x10)
-        x20 = torch.randn((k, num_samples), generator=gen)
+        x20 = torch.randn((k, num_samples), generator=gen, dtype=torch.float64)
         x2 = torch.Tensor(A2 @ x20)
         # compute densities for each
         p1 = -torch.sum(x10**2, 0) / 2 - logdet1
